♨

+++ upgrade.package.json.sh
@@ -9,9 +9,11 @@ for p in $(fd package.json \
   -E .git \
   -E node_modules \
   -aIHt f); do
-  cmd="cd $(dirname $p) && ncu -u && ni"
-  echo -e "$GREEN> $cmd$NC"
-  bash -c "$cmd"
+  if [ ! grep -q -E '/api/' $p ]; then
+    cmd="cd $(dirname $p) && ncu -u && ni"
+    echo -e "$GREEN> $cmd$NC"
+    bash -c "$cmd"
+  fi
 done
 
 ./git.sync.sh
